Move ImportDataView into dedicated folder
