+++ source/darefl/importdataview/CMakeLists.txt
@@ -1,10 +1,12 @@
 target_sources(${library_name} PRIVATE
+    datarowstrategy.cpp
+    datarowstrategy.h
     dataselectionmodel.cpp
     dataselectionmodel.h
     dataviewmodel.cpp
     dataviewmodel.h
-    datarowstrategy.cpp
-    datarowstrategy.h
     importdataeditor.cpp
     importdataeditor.h
+    importdataview.cpp
+    importdataview.h
 )
