+++ source/darefl/mainwindow/mainwindow.cpp
@@ -13,7 +13,7 @@
 #include <QMenuBar>
 #include <QSettings>
 #include <darefl/mainwindow/actionmanager.h>
-#include <darefl/mainwindow/importdataview.h>
+#include <darefl/importdataview/importdataview.h>
 #include <darefl/mainwindow/mainbarwidget.h>
 #include <darefl/mainwindow/mainwindow.h>
 #include <darefl/mainwindow/refldockwindow.h>
